updated yolo world and test metadata
