+++ tests/experiments_metadata.json
@@ -1,40 +1,12 @@
 {
     "known_failing_experiments": {
-        "neural-networks/facial-detection/age-gender": {
-            "reason": "Usage of MessageGroup not supported on RVC2 yet",
-            "platform": [
-                "rvc2"
-            ]
-        },
-        "neural-networks/facial-detection/fatigue-detection": {
-            "reason": "Usage of MessageGroup not supported on RVC2 yet",
-            "platform": [
-                "rvc2"
-            ]
-        },
-        "neural-networks/facial-detection/gaze-estimation": {
-            "reason": "Usage of MessageGroup not supported on RVC2 yet",
-            "platform": [
-                "rvc2"
-            ]
-        },
-        "neural-networks/reidentification/human-reidentification": {
-            "reason": "Usage of MessageGroup not supported on RVC2 yet",
-            "platform": [
-                "rvc2"
-            ]
-        },
-        "neural-networks/facial-detection/head-posture-detection": {
-            "reason": "luxonis/yunet:640x640 not available anymore, should be swaped with luxonis/yunet:640x480 + some edits in the pipeline",
-            "platform": "all"
-        },
         "neural-networks/speech-recognition/whisper-tiny-en": {
-            "reason": "Can't run without arguments, needs to be updated",
+            "reason": "Complex example, needs to be updated",
             "platform": "all"
         },
         "neural-networks/object-detection/yolo-world": {
-            "reason": "Can't run without arguments, needs to be updated",
-            "platform": "all"
+            "reason": "Complex example, works only on RVC4",
+            "platform": "rvc4"
         },
         "neural-networks/object-detection/thermal-detection": {
             "reason": "Needs thermal camera",
@@ -44,50 +16,32 @@
             "reason": "Failed to connect to Rerun server",
             "platform": "all"
         },
-        "neural-networks/ocr/general-ocr": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
-            "platform": [
-                "rvc4"
-            ]
-        },
         "neural-networks/object-tracking/kalman": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
+            "reason": "Pipeline node with name: 'ObjectTracker' doesn't exist",
             "platform": [
                 "rvc4"
             ]
         },
         "neural-networks/object-tracking/collision-avoidance": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
+            "reason": "Pipeline node with name: 'ObjectTracker' doesn't exist",
             "platform": [
                 "rvc4"
             ]
         },
         "neural-networks/object-tracking/people-tracker": {
-            "reason": "ObjectTracker node not available on RVC4",
+            "reason": "Pipeline node with name: 'ObjectTracker' doesn't exist",
             "platform": [
                 "rvc4"
             ]
         },
         "neural-networks/counting/cumulative-object-counting": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
+            "reason": "Pipeline node with name: 'ObjectTracker' doesn't exist",
             "platform": [
                 "rvc4"
             ]
         },
         "neural-networks/counting/depth-people-counting": {
             "reason": "Pipeline node with name: 'ObjectTracker' doesn't exist",
-            "platform": [
-                "rvc4"
-            ]
-        },
-        "neural-networks/object-detection/fire-detection": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
-            "platform": [
-                "rvc4"
-            ]
-        },
-        "neural-networks/segmentation/depth-crop": {
-            "reason": "Loaded model is for RVC2, but the device is RVC4",
             "platform": [
                 "rvc4"
             ]
@@ -159,41 +113,5 @@
         "The issued warnings are orientative, based on optimal settings for a single network, if multiple networks are running in parallel the optimal settings may vary",
         "Did not get the input image sizes from the imageIn input. Defaulting to 416 x 416",
         "Network compiled for 8 shaves, maximum available"
-    ],
-    "experimental_subset": [
-        "neural-networks/counting/crowdcounting",
-        "neural-networks/counting/people-counter",
-        "neural-networks/facial-detection/face-mask-detection",
-        "neural-networks/generic-example",
-        "neural-networks/object-tracking/people-tracker",
-        "neural-networks/pose-estimation/hand-pose",
-        "neural-networks/3D-detection/objectron",
-        "neural-networks/depth-estimation/crestereo-stereo-matching",
-        "neural-networks/facial-detection/age-gender",
-        "neural-networks/facial-detection/blur-faces",
-        "neural-networks/facial-detection/emotion-recognition",
-        "neural-networks/feature-detection/xfeat",
-        "neural-networks/object-detection/human-machine-safety",
-        "neural-networks/object-detection/social-distancing",
-        "neural-networks/object-detection/spatial-detections",
-        "neural-networks/object-detection/text-blur",
-        "neural-networks/object-detection/thermal-detection",
-        "neural-networks/object-detection/yolo-host-decoding",
-        "neural-networks/object-detection/yolo-p",
-        "neural-networks/object-detection/yolo-world",
-        "neural-networks/object-tracking/deepsort-tracking",
-        "neural-networks/pose-estimation/animal-pose",
-        "neural-networks/reidentification/human-reidentification",
-        "neural-networks/segmentation/yolov8-instance-segmentation",
-        "streaming/mjpeg-streaming",
-        "streaming/on-device-encoding",
-        "streaming/rtsp-streaming",
-        "streaming/webrtc-streaming",
-        "tutorials/camera-demo",
-        "tutorials/custom-models",
-        "tutorials/display-detections",
-        "tutorials/full-fov-nn",
-        "tutorials/play-encoded-stream",
-        "tutorials/qr-with-tiling"
     ]
 }
